내 오디오stream 컴포넌트 화 해서 리렌더링 해결
